test(config): keep app-config free of test helpers
